requirements.txt
distro==1.8.0
JPype1 @ git+https://github.com/jpype-project/jpype.git@0c50910dbdee075549fbbe4595b9c117db04dbf6
numpy==1.26.2
packaging==23.2
pandas==2.1.4
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tabula-py==2.9.0
tzdata==2023.3
